+++ pnet-data-api-java/src/main/java/pnet/data/api/todo/TodoState.java
@@ -10,42 +10,54 @@ public enum TodoState
     /**
      * The item waits to be approved.
      */
-    PENDING,
+    PENDING(TodoPhase.IN_APPROVAL),
 
     /**
      * The item is waiting for it's execution.
      */
-    PLANNED,
+    PLANNED(TodoPhase.IN_PROGRESS),
 
     /**
      * The item get's executed right now.
      */
-    EXECUTING,
+    EXECUTING(TodoPhase.IN_PROGRESS),
 
     /**
      * The item was executed and has been finished.
      */
-    DONE,
+    DONE(TodoPhase.COMPLETED),
 
     /**
      * The item has been acknowledged, which is a special form of done.
      */
-    ACKNOWLEDGED,
+    ACKNOWLEDGED(TodoPhase.COMPLETED),
 
     /**
      * The execution has failed.
      */
-    FAILED,
+    FAILED(TodoPhase.COMPLETED),
 
     /**
      * The item has been rejected.
      */
-    REJECTED,
+    REJECTED(TodoPhase.COMPLETED),
 
     /**
      * The item has been ignored, which is a special form of rejected.
      */
-    IGNORED;
+    IGNORED(TodoPhase.COMPLETED);
+
+    private final TodoPhase phase;
+
+    TodoState(TodoPhase phase)
+    {
+        this.phase = phase;
+    }
+
+    public TodoPhase getPhase()
+    {
+        return phase;
+    }
 
     public static boolean isRejected(TodoState state)
     {