+++ CHANGELOG.md
@@ -2,6 +2,7 @@ pnet-data-api 1.8.2
 ===================
 
 * The id of the person was renamed to personId (the old field is deprecated)
+* Added the WithCompanyId interface, where applicable.
 
 pnet-data-api 1.8.1
 ===================